# api/openapi/admin-v1.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "Waiting Room Admin API",
    "version": "0.1.0",
    "description": "Private Admin authentication and persisted configuration draft API with a local Backoffice UI. Draft saves do not publish or activate queue runtime. Local Docker persistence is under development; this is not a production/Beta readiness claim. Pinned Origin required. Bootstrap requires a separate literal-loopback TLS listener. Challenges are not sessions. Other planned operations remain contract-only unless explicitly verified.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://admin.waiting.example.test/api/admin/v1",
      "description": "Private admin host"
    }
  ],
...
  "security": [
    {
      "adminSession": []
    }
  ],
  "paths": {
    "/bootstrap": {
      "post": {
        "operationId": "bootstrap",
        "summary": "Create the first administrator",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bootstrapToken": []
          }
        ],
        "x-required-roles": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "description": "Must exactly match the configured HTTPS admin origin.",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "X-WR-Auth",
            "in": "header",
            "required": true,
            "description": "Non-simple pre-auth request marker; not an identity credential.",
            "schema": {
              "const": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string",
                  "minLength": 1
                },
                "description": "Only full authentication issues __Host-wrs; Secure; HttpOnly; SameSite=Strict; Path=/."
              }
            }
          },
          "400": {
            "description": "Invalid input.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "403": {
            "description": "Permission or TOTP policy denied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "405": {
            "description": "POST required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Allow": {
                "schema": {
                  "const": "POST"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 4096 bytes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "415": {
            "description": "Only application/json (optional UTF-8 charset), no content encoding.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          },
          "429": {
            "description": "Shared source or installation auth HTTP budget exhausted (AUTH_RATE_LIMITED).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "const": "60"
                }
              }
            }
          },
          "503": {
            "description": "Authentication capacity, deadline support or storage unavailable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bootstrap"
              }
            }
          }
        },
        "description": "Local direct TLS adapter: pinned Host and exact Origin, X-WR-Auth: 1, application/json, 4096-byte body cap. No Cookie or X-CSRF-Token; log out before authentication. No CORS or proxy trust. Production listener and UI remain pending. Separate bootstrap handler only; literal loopback HTTPS origin, loopback local socket and peer required. X-Bootstrap-Token only, no Authorization. Operator token issuance is not an HTTP API. DB TOTP policy is authoritative; wizard policy selection pending."
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-wrs"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "challengeToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Scoped one-time challenge; five-minute TTL"
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Bootstrap-Token",
        "description": "Loopback installation only. Permanently disabled after bootstrap."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "IDEMPOTENCY_CONFLICT",
              "ROUTE_CONFLICT",
              "EVENT_OVERLAP",
              "TOTP_ENROLLMENT_REQUIRED",
              "TOTP_INVALID_OR_REPLAYED",
              "TOTP_POLICY_LOCKED",
              "REVISION_MISMATCH",
              "UNSUPPORTED_CAPABILITY",
              "PRECONDITION_REQUIRED",
              "API_RATE_LIMITED",
              "QUEUE_UNAVAILABLE",
              "AUTH_UNAVAILABLE",
              "AUTH_RATE_LIMITED",
              "CONTROL_UNAVAILABLE",
              "COMMAND_CAPACITY_EXCEEDED",
              "INVALID_CONFIG"
            ]
          },
          "requestId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "requestId"
        ]
      },
      "CurrentSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["userId", "role", "mfaVerified", "totpEnrolled", "capabilityVersion", "capabilities", "createdAt", "lastSeenAt", "idleExpiresAt", "absoluteExpiresAt"],
        "properties": {
          "userId": {"type": "string", "minLength": 1},
          "role": {"enum": ["admin", "operator", "viewer"], "description": "Core capability role IDs, not legacy User display labels."},
          "mfaVerified": {"type": "boolean"},
          "totpEnrolled": {"type": "boolean"},
          "capabilityVersion": {"const": 1},
          "capabilities": {
            "type": "array",
            "items": {
              "type": "object", "additionalProperties": false,
              "required": ["action", "requiresReauthentication"],
              "properties": {
                "action": {"type": "string", "minLength": 1},
                "requiresReauthentication": {"type": "boolean"}
              }
            }
          },
          "createdAt": {"type": "string", "format": "date-time"},
          "lastSeenAt": {"type": "string", "format": "date-time"},
          "idleExpiresAt": {"type": "string", "format": "date-time"},
          "absoluteExpiresAt": {"type": "string", "format": "date-time"}
        }
      },
...
      "LoginResult": {
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "state": {
                "const": "authenticated"
              },
              "csrfToken": {
                "type": "string",
                "minLength": 43,
                "maxLength": 43,
                "pattern": "^[A-Za-z0-9_-]{43}$"
              },
              "session": {
                "$ref": "#/components/schemas/CurrentSession"
              }
            },
            "required": [
              "state",
              "session",
              "csrfToken"
            ]
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "state": {
                "enum": [
                  "totp_required",
                  "enrollment_required"
                ]
              },
              "challengeToken": {
                "type": "string",
                "minLength": 43,
                "maxLength": 43,
                "pattern": "^[A-Za-z0-9_-]{43}$"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "state",
              "challengeToken",
              "expiresAt"
            ]
          }
        ]
      },
...
      "Bootstrap": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]{1,128}$"
          },
          "password": {
            "type": "string",
            "minLength": 15,
            "maxLength": 1024,
            "writeOnly": true,
            "description": "15 Unicode code points minimum, 1024 UTF-8 bytes maximum. No normalization."
          }
        },
        "required": [
          "username",
          "password"
        ],
        "description": "DB TOTP policy only. Client role, secret and totpEnabled rejected; wizard policy write pending."
      }
    }
  }
}